Feat: Add Modules logic
